Remove redundant code, add documentation to functions, and delete some empty print statements.
The print statements are there specifically to split up the different parts of the output. However, I decided that it just makes the output feel more bulky than necesssary and I prefer it to be more compact to minimize the amount of scrolling up and down. Also, I switched to writing the commits imperatively so that I'm practicing following best practices and also it will make more sense when/if I need to revert to a previous commit.