Split main() into other smaller functions, update function documentation, change some of the code to be more maintainable in the future.
This is a pretty hefty commit, most of the file was modified in one way or another. Originally, I was thinking about rewriting bits to practice utilizing classes, but I'm not sure if it's necessary or would benefit the project in any way. This commit should allow the script to be maintained more easily.